requirements.txt

requests
py-solana
pyTelegramBotAPI
Flask
